fix list view delete option.

+++ todo-frontend/src/components/listView.js
@@ -1,35 +1,38 @@
 import React, { Component } from "react";
 import { Table, Button, Row, Glyphicon } from "react-bootstrap";
 import swal from "sweetalert2";
+import moment from "moment";
 
 class ListView extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      todos: []
-    };
-  }
-
   componentDidMount() {
     fetch("http://localhost:4000/todos")
       .then(res => res.json())
-      .then(todos => this.setState({ todos }));
+      .then(todos => this.props.updateTodos(todos));
   }
 
   delTask = (id, index) => {
-    fetch("http://localhost:4000/todos/" + id, {
-      method: "delete"
-    })
-      .then(res => res.json())
-      .then(res => {
-        if(res.status){
-          const {todos} = this.state
-          todos.splice(index,1)
-          this.setState(todos)
-          swal("Todo! " + id, "Todo Deleted!", "error");
-        }
-      })
-      .catch(err => console.log(err));
+    swal({
+      title: "Are you sure ?",
+      text: "Todo Deleted!",
+      type: "error",
+      showCancelButton: true,
+      cancelButtonColor: "#d33"
+    }).then(d => {
+      if (d.value) {
+        fetch("http://localhost:4000/todos/" + id, {
+          method: "delete"
+        })
+          .then(res => res.json())
+          .then(res => {
+            if (res.status) {
+              const { todos } = this.props;
+              todos.splice(index, 1);
+              this.props.updateTodos(todos);
+            }
+          })
+          .catch(err => console.log(err));
+      }
+    });
   };
 
   updateTask = (title, content, index) => {
@@ -49,13 +52,12 @@ class ListView extends Component {
         });
       }
     }).then(function(result) {
-      console.log({ result });
       swal("Todo!", "Your Todo Has Been Updated!", "success");
     });
   };
 
   render() {
-    const { todos } = this.state;
+    const { todos } = this.props;
     return (
       <div>
         <Row>
@@ -82,12 +84,13 @@ class ListView extends Component {
           </thead>
           <tbody>
             {todos.map((d, index) => {
+              const date = moment(d.date);
               return (
                 <tr key={index.toString()}>
                   <td>{index + 1}</td>
                   <td>{d.title}</td>
                   <td>{d.content}</td>
-                  <td>{d.date}</td>
+                  <td>{date.format("L")}</td>
                   <td>
                     <Button
                       bsStyle="primary"
